bet_maker/pyproject.toml
[tool.poetry]
name = "bet_maker"
version = "0.1.0"
description = ""
authors = ["Артём Поташов <artem.potashov@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
tortoise-orm = "^0.20.0"
uvicorn = "^0.25.0"
asyncpg = "^0.29.0"
aiokafka = "^0.10.0"
httpx = "^0.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"


[tool.aerich]
tortoise_orm = "tortoise_conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
